backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.7.1
python-multipart==0.0.20
email-validator==2.2.0
minio==7.2.13
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.6.23
requests==2.32.3
gunicorn==23.0.0
